requirements.txt
python-dotenv
pandas
pymongo
numpy
certifi
fastapi
pymongo[srv]==3.12
scikit-learn
mlflow==2.11.3
pyaml
dagshub

uvicorn
python-multipart
requests
beautifulsoup4
python-whois
dnspython


#-e 